src/requirements.txt
requests
pandas
torch
transformers
openai
python-dotenv
logging
numpy
tqdm
scikit-learn
torch
transformers==4.31.0
sentence-transformers==2.2.2
matplotlib
optuna
boto3
relativedelta
shap